Perform checkouts only at the workflow level
Ok, the content wasn't there because like an idiot I wasn't running actions/checkout. However, this was a good issue to run into, because it highlighted that my checkout calls were all over the place -- sometimes in the actions, sometimes in the workflows. After thinking a bit about it, I'm opting to *only* put checkout calls in the workflows themselves. This adds an extra step before using most of the actions in downstream workflows, but it prevents wasted time with repeated checkouts and prevents security holes where people accidentally use one of our actions in a pr-target trigger.

We may want to think about putting the doc env commit only in pushes to main and not push or pull, but it's inexpensive and most of the time should be a zero-diff commit, so for now I err on the side of doing it too much.