Rewrite cached-miniforge to avoid `mamba update`
There is an awkward edge case when the update env explicitly specifies a range of python versions that this can override the python version specified as an argument to `conda-incubator/setup-miniconda`; although the update approach is suggested in that action's docs for caching an env, the approach here is safer. A new input variable is introduced to store the location of the env, but the old `miniforge-activate-environment` can still be specified for backwards compatibility (although what happens under the hood has changed).