Manage coveralls and codacy dependencies by boolean
So that the upstream callers can turn them on or off. If users want some non-standard version, they already have the freedom to add these via the `env-files` argument, we anyhow only really needed these input fields to access the default stored env file, so a boolean is completely ok!

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>