[patch] Use codecov (#146)
* Target branch

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Include codecov as an option for coverage

And make the language more generic now that there are two paths

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Break apart the coveralls and codacy envs

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Manage coveralls and codacy dependencies by boolean

So that the upstream callers can turn them on or off. If users want some non-standard version, they already have the freedom to add these via the `env-files` argument, we anyhow only really needed these input fields to access the default stored env file, so a boolean is completely ok!

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Reinstate the removed args

And change the default of the new args to maintain backwards compatibility

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Combine input args

To match the original for compatiblity

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Fill in final target

I must have overwritten this replacement somehow?

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Normalize path separators

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Move slash

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Print paths

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Normalize the standard env path too

This is the one which actually provides coverage

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Actually forward the codecov input

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

* Target the next release

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>

---------

Signed-off-by: liamhuber <liamhuber@greyhavensolutions.com>